Move mandataris publication status logic to publication-status-selector
